+++ CHANGELOG.md
@@ -2,16 +2,26 @@ Next
 ====
 
 * Your contribution here.
+
+0.2.0
+====
+
 * Fix file-handle leak - [@justfalter](https://github.com/justfalter)
-[#115](https://github.com/cikl/threatinator/issues/56): Add missing require for 'set' - [@justfalter](https://github.com/justfalter)
-* [#121](https://github.com/cikl/threatinator/issues/121) Add threatinator list 'json' output format. [@justfalter](https://github.com/justfalter)
 * BREAKING CHANGE: Event#ipv4s -> now an Ipv4ObservableCollection that consists on Ipv4Observable objects. 
 * Normalize and validate ipv4 addresses.
 * Add URL support [@justfalter](https://github.com/justfalter)
 * Re-implement Event model, get rid of PropertyDefiner [@justfalter](https://github.com/justfalter)
 * Disable decoding when testing the parsing of a single record. [@justfalter](https://github.com/justfalter)
-* Added Danger bruteforce feed. - [@pierre427](https://github.com/pierre427)
-* Added trustedsec feed. - [@pierre427](https://github.com/pierre427)
+
+0.1.6
+====
+
+* [#121](https://github.com/cikl/threatinator/issues/121) Add threatinator list 'json' output format. [@justfalter](https://github.com/justfalter)
+
+0.1.5
+====
+
+* [#115](https://github.com/cikl/threatinator/issues/56): Add missing require for 'set' - [@justfalter](https://github.com/justfalter)
 
 0.1.2
 ====